requirements.txt
Django==2.1.2
trello==0.9.1
djangorestframework==3.9.0
